+++ pytsite/google/res/assets/js/pytsite-google.js
@@ -1,5 +1,8 @@
 pytsite.google = {
+    ready: false,
+
     initCallback: function() {
         $(window).trigger('pytsite.google.ready');
+        pytsite.google.ready = true;
     }
 };